requirements.txt
torch==2.3.0
torchvision==0.14.1
numpy==1.26.4
tqdm==4.66.4 
scipy==0.16.0
Pillow==10.3.0
ml-collections==0.1.1